#13: Move common populator test constants and helper classes into common MessagePopulatorTests utility class.
